fixed OIDC login
